Fix and refactor
