+++ saleor/graphql/app/mutations/app_fetch_manifest.py
@@ -5,12 +5,8 @@ import requests
 from django.core.exceptions import ValidationError
 
 from ....app.error_codes import AppErrorCode
-from ....app.installation_utils import REQUEST_TIMEOUT
-from ....app.manifest_validations import (
-    clean_manifest_data,
-    clean_manifest_url,
-    fetch_brand_data,
-)
+from ....app.installation_utils import REQUEST_TIMEOUT, fetch_brand_data
+from ....app.manifest_validations import clean_manifest_data, clean_manifest_url
 from ....permission.enums import AppPermission
 from ...core import types as grapqhl_types
 from ...core.doc_category import DOC_CATEGORY_APPS
